fixes bug #4.

+++ src/main/java/sds/classfile/attributes/annotation/ParameterAnnotations.java
@@ -6,7 +6,7 @@ import java.io.RandomAccessFile;
 /**
  * This class is for annotations table.<br>
  * {@link RuntimeVisibleParameterAnnotations <code>RuntimeVisibleParameterAnnotations</code>}
- * and 
+ * and
  * {@link RuntimeInvisibleParameterAnnotations <code>RuntimeInvisibleParameterAnnotations</code>}
  * have item.
  * @author inagaki
@@ -15,7 +15,7 @@ public class ParameterAnnotations {
 	private Annotation[] annotations;
 
 	ParameterAnnotations(RandomAccessFile raf) throws IOException, ElementValueException {
-		this.annotations = new Annotation[raf.readByte()];
+		this.annotations = new Annotation[raf.readShort()];
 		for(int i = 0; i < annotations.length; i++) {
 			annotations[i] = new Annotation(raf);
 		}
